requirements.txt
python3-ghostscript==0.5.0
